Make the webinstaller optional
The new default is to load in all necessary SQL files through
koha-mysql. The user can choose to run the webinstaller by
setting skip_webinstaller="0" in the config file. Which SQL
files will be loaded can be set with installer_lang and
installer_marcflavour.

An accompanying change is that koha-sites.cfg is just provided
as a sample, and the user has to make a copy that can be edited.
This makes it possible to set language and marcflavour here as
well.

Fixes #1.