Clone the community repo if SYNC_REPO is not defined
The original kohadevbox clones the community repo if the SYNC_REPO
environment variable is not defined. This patch introduces a check
for this ENV variable in the Vagrantfile, and injects its value
into ansible.extra_vars (sync_repo var).

It also adds the (conditional to the value of sync_repo) clone
action to gitify.yml.

I took the liberty to also add the SKIP_WEBINSTALLER env variable
using the same schema, but it is not yet used to match the original
kohadevbox's functionality.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>