Use NFS instead of vboxfs for vagrant-cachier
Because of its bad performance and the fact that vboxfs requires contrib or
non-free packages, the extension pack has been removed from Jessie's official
image (as of 8.2.2). People using vagrant-cachier won't be able to load kohadevbox
if this is not moved to NFS or another method.

This patch makes vagrant-cachier use NFS. If you can't use NFs, don't use vagrant-cachier
until we find another solution.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>