Add the install steps for ElasticSearch
This patch introduces a task for installing ES inside kohadevbox. It is controlled by
the new 'elasticsearch' var (in vars/defaults.yml and vars/user.yml) and the (handy) KOHA_ELASTICSEARCH
en var. So you can provision  ES by issuing:

  $ KOHA_ELASTICSEARCH vagrant provision jessie

Right now it only installs ES with it's dependencies and launches it (it is configured to do so
on startup).

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>