Allow using custom deb repositories
This refactoring allows the user to add its custom repository. This is specially
useful when testing package-specific stuff and the developer has provided a repository
with packages for its changes.

It allows pinning the custom repo so koha-common is picked from there. All this behaviour
is controled by configuration variables in vars/{defaults|user}.yml. Please make your
changes in user.yml.

It comes with Theke's repo for the master packages as they are still outdated in the
official repo. This will be reverted when master packages are officialy built.

Signed-off-by: Tomas Cohen Arazi <tomascohen@unc.edu.ar>