Issue #135: Initial support for ES 2.4
This patch introduces a new task file for handling ES 2.4. This
way we can keep the 1.7 one and work on this without messing with tons
of conditionals.

It introduces the requirement for Oracle Java 8 (officially required by ES)
but this could be made optional on a followup.

Right now it only installs ES 2.4 and Java 8.

TODO: have dependencies installed (Catmandu 1.x and friends)

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>