Issue #176: Reorganize tasks
This patch reorganizes tasks. It adds more granularity to handle the way
instance creation using the repository's scripts.

Instance creation and gitifying is done on a separate task so it is
easier to spot where and when things happen during the playbook run.

Touched files are moved to the newer syntax.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>