Issue #203: Use Vagrant capabilities to make Ansible available on guest OS
As proposed by @minusdavid, Vagrant provides a native way of installing Ansible
on the guest OS. We should use it instead of our install-ansible.sh script.

This patch implements it.

For the record, I chose using pip for the install process, to have more control on
the chosen version for the install process (it will otherwise install an old one in Stretch
and Jessie, etc).

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>